fixed 3 + new 4

+++ lab3/lab3.2/main.c
@@ -6,17 +6,19 @@
 #include <unistd.h>
 
 
-void CreateDir(const char* dir_name){
-    if(mkdir(dir_name, 0777) == 0){
-        printf("dir created successfully\n");
-    } else {
+int CreateDir(const char* dir_name){
+    if(mkdir(dir_name, 0777) != 0){
         perror("error in creating dir\n");
+        return 1;
     }
+    return 0;
 }
+
 void ListDir(const char* dir_name){
     DIR* dir = opendir(dir_name);
     if(dir == NULL){
         perror("error in opening dir\n");
+        return;
     }
     struct dirent* cur_file;
     while((cur_file = readdir(dir)) != NULL){
@@ -26,36 +28,37 @@ void ListDir(const char* dir_name){
 }
 
 void DeleteDir(const char* dir_name){
-    if(rmdir(dir_name) == 0){
-        printf("dir deleted successfully\n");
-    } else {
+    if(rmdir(dir_name) != 0){
         perror("error in deleting dir\n");
+        return;
     }
 }
 
-void CreateFile(const char* file_name){
+int CreateFile(const char* file_name){
     FILE* file = fopen(file_name, "w+");
-    if(file != NULL){
-        printf("file created successfully\n");
-    } else {
+    if(file == NULL){
         perror("error in creating file\n");
+        return 1;
     }
-    if(fclose(file) != 0){
+    if(fclose(file) == 0){
         perror("error in closing file");
+        return 1;
     }
+    return 0;
 }
 
 void PrintFile(const char* file_name){
     FILE* file = fopen(file_name, "rb");
     if(file == NULL){
         perror("error in opening file");
+        return;
     }
     fseek(file, 0, SEEK_END);
     long file_size = ftell(file);
     fseek(file, 0, SEEK_SET);
     char* buffer = (char*) malloc(sizeof(char) * file_size);
     size_t bytes_read = fread(buffer, 1, file_size, file);
-    if (bytes_read != file_size) {
+    if(bytes_read != file_size){
         perror("error in reading file\n");
         free(buffer);
         if(fclose(file) != 0){
@@ -69,23 +72,23 @@ void PrintFile(const char* file_name){
     free(buffer);
     if(fclose(file) != 0){
         perror("error in closing file");
+        return;
     }
 }
 
 void DeleteFile(const char* file_name){
-    if(remove(file_name) == 0){
-        printf("file deleted successfully\n");
-    } else {
+    if(remove(file_name) != 0){
         perror("error in deleting file\n");
+        return;
     }
 }
 
-void CreateSymlink(char* file_name, char* link_name){
-    if(symlink(file_name, link_name) == 0){
-        printf("symlink created successfully\n");
-    } else {
+int CreateSymlink(char* file_name, char* link_name){
+    if(symlink(file_name, link_name) != 0){
         perror("error in creating symlink\n");
+        return 1;
     }
+    return 0;
 }
 
 void PrintSymlink(const char* link_name){
@@ -104,33 +107,31 @@ void PrintFileFromSymlink(const char* link_name){
     size_t bytes_read = readlink(link_name, buffer, sizeof(buffer));
     if(bytes_read == -1){
         perror("error in reading symlink\n");
-    } else {
-        buffer[bytes_read] = '\0';
-        PrintFile(buffer);
+        return;
     }
+    buffer[bytes_read] = '\0';
+    PrintFile(buffer);
 }
 
 void DeleteSymlink(const char* link_name){
     if(unlink(link_name) == -1){
         perror("error in deleting symlink\n");
-    } else {
-        printf("symlink deleted successfully\n");
+        return;
     }
 }
 
-void CreateHardlink(const char* file_name,const char* link_name){
+int CreateHardlink(const char* file_name, const char* link_name){
     if(link(file_name, link_name) == -1){
         perror("error in creating hardlink\n");
-    } else {
-        printf("hardlink created successfully\n");
+        return 1;
     }
+    return 0;
 }
 
 void DeleteHardlink(char* link_name){
     if(unlink(link_name) == -1){
         perror("error in deleting hardlink\n");
-    } else {
-        printf("hardlink deleted successfully\n");
+        return;
     }
 }
 
@@ -138,27 +139,28 @@ void PrintInfo(char* file_name){
     struct stat file_stat;
     if(stat(file_name, &file_stat) != 0){
         perror("error in getting info");
+        return;
     }
     char* mode = (char*) malloc(sizeof(char) * 11);
-    for(int i = 0; i < 9; i++) {
+    for(int i = 0; i < 9; i++){
         mode[i] = '-';
     }
     mode[10] = '\0';
     
-    if (S_ISDIR(file_stat.st_mode)) mode[0] = 'd';
-    else if (S_ISLNK(file_stat.st_mode)) mode[0] = 'l';
+    if(S_ISDIR(file_stat.st_mode)) mode[0] = 'd';
+    else if(S_ISLNK(file_stat.st_mode)) mode[0] = 'l';
     
-    if ((file_stat.st_mode & S_IRUSR) != 0) mode[1] = 'r';
-    if ((file_stat.st_mode & S_IWUSR) != 0) mode[2] = 'w';
-    if ((file_stat.st_mode & S_IXUSR) != 0) mode[3] = 'x';
-    if ((file_stat.st_mode & S_IRGRP) != 0) mode[4] = 'r';
-    if ((file_stat.st_mode & S_IWGRP) != 0) mode[5] = 'w';
-    if ((file_stat.st_mode & S_IXGRP) != 0) mode[6] = 'x';
-    if ((file_stat.st_mode & S_IROTH) != 0) mode[7] = 'r';
-    if ((file_stat.st_mode & S_IWOTH) != 0) mode[8] = 'w';
-    if ((file_stat.st_mode & S_IXOTH) != 0) mode[9] = 'x';
+    if((file_stat.st_mode & S_IRUSR) != 0) mode[1] = 'r';
+    if((file_stat.st_mode & S_IWUSR) != 0) mode[2] = 'w';
+    if((file_stat.st_mode & S_IXUSR) != 0) mode[3] = 'x';
+    if((file_stat.st_mode & S_IRGRP) != 0) mode[4] = 'r';
+    if((file_stat.st_mode & S_IWGRP) != 0) mode[5] = 'w';
+    if((file_stat.st_mode & S_IXGRP) != 0) mode[6] = 'x';
+    if((file_stat.st_mode & S_IROTH) != 0) mode[7] = 'r';
+    if((file_stat.st_mode & S_IWOTH) != 0) mode[8] = 'w';
+    if((file_stat.st_mode & S_IXOTH) != 0) mode[9] = 'x';
     
-    for(int i = 0; i < 9; i++) {
+    for(int i = 0; i < 9; i++){
         printf("%c", mode[i]);
     }
     printf("\n");
@@ -167,25 +169,26 @@ void PrintInfo(char* file_name){
 void ChangeMode(char* file_name, char* new_mode){
     mode_t mode = 0;
     for(int i = 0; i < strlen(new_mode); i++){
-        if(new_mode[i] == 'r' && i==1) mode |= S_IRUSR;
-        else if(new_mode[i] == 'r' && i==4) mode |= S_IRGRP;
-        else if(new_mode[i] == 'r' && i==7) mode |= S_IROTH;
+        if(new_mode[i] == 'r' && i == 1) mode |= S_IRUSR;
+        else if(new_mode[i] == 'r' && i == 4) mode |= S_IRGRP;
+        else if(new_mode[i] == 'r' && i == 7) mode |= S_IROTH;
         
-        else if(new_mode[i] == 'w' && i==2) mode |= S_IWUSR;
-        else if(new_mode[i] == 'w' && i==5) mode |= S_IWGRP;
-        else if(new_mode[i] == 'w' && i==8) mode |= S_IWOTH;
+        else if(new_mode[i] == 'w' && i == 2) mode |= S_IWUSR;
+        else if(new_mode[i] == 'w' && i == 5) mode |= S_IWGRP;
+        else if(new_mode[i] == 'w' && i == 8) mode |= S_IWOTH;
         
-        else if(new_mode[i] == 'x' && i==3) mode |= S_IXUSR;
-        else if(new_mode[i] == 'x' && i==6) mode |= S_IXGRP;
-        else if(new_mode[i] == 'x' && i==9) mode |= S_IXOTH;
+        else if(new_mode[i] == 'x' && i == 3) mode |= S_IXUSR;
+        else if(new_mode[i] == 'x' && i == 6) mode |= S_IXGRP;
+        else if(new_mode[i] == 'x' && i == 9) mode |= S_IXOTH;
         
         else if(new_mode[i] == '-') continue;
         else {
             perror("incorrect mode symbol");
         }
     }
-    if (chmod(file_name, mode) == -1) {
+    if(chmod(file_name, mode) == -1){
         perror("error in chanding mode");
+        return;
     }
 }
 
@@ -196,26 +199,26 @@ int main(int argc, char* argv[]){
     }
     
     char* dir_name = argv[1];
-    CreateDir(dir_name);
+    if(CreateDir(dir_name) == 1) return 1;
     ListDir(dir_name);
     DeleteDir(dir_name);
     
     char* file_name = argv[2];
-    CreateFile(file_name);
+    if(CreateFile(file_name) == 1) return 1;
     PrintInfo(file_name);
     char* new_mode = "-rwxrwxrwx";
     ChangeMode(file_name, new_mode);
-
+    
     PrintInfo(file_name);
     PrintFile(file_name);
     
     char* symlink_name = argv[3];
-    CreateSymlink(file_name, symlink_name);
+    if(CreateSymlink(file_name, symlink_name) == 1) return 1;
     PrintSymlink(symlink_name);
     DeleteSymlink(symlink_name);
     
     char* hardlink_name = argv[4];
-    CreateHardlink(file_name, hardlink_name);
+    if(CreateHardlink(file_name, hardlink_name) == 1) return 1;
     DeleteHardlink(hardlink_name);
     
     DeleteFile(file_name);
